refactor: modularize Coursify reader and sync public/internal UIs
- Extracted monolithic logic into shared hooks: `useCourseReader`, `useTableOfContents`, and `useReaderUI`.
- Created modular UI components in `src/components/coursify/reader/`: `ReaderHeader`, `ReaderSidebar`, `CourseOverview`, `MarkdownRenderer`, `TableOfContents`, and `ReaderNavigation`.
- Refactored `src/app/coursify/[slug]/page.js` (Public Reader) to use new modular architecture.
- Refactored `src/app/apps/coursify/[id]/page.js` (Internal Reader) to use new modular architecture while preserving administrative features (Planning workspace, Edit mode, Section management).
- Ensured UI consistency and feature parity between public and internal reader views.
