+++ src/app/coursify/[slug]/page.js
@@ -1,27 +1,20 @@
 'use client';
 
-import { useState, useEffect, useCallback, use, useRef } from 'react';
+import { useEffect, use, useRef } from 'react';
 import { useRouter } from 'next/navigation';
 import { Pacifico, Nunito, Lora } from 'next/font/google';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import rehypeRaw from 'rehype-raw';
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { oneDark } from 'react-syntax-highlighter/dist/esm/styles/prism';
-import {
-  ArrowLeft,
-  BookOpen,
-  Clock,
-  Layers,
-  Tag,
-  ExternalLink,
-  Menu,
-  X,
-  ChevronRight,
-  CheckCircle2,
-  Check,
-  List,
-} from 'lucide-react';
+import { ExternalLink, BookOpen } from 'lucide-react';
+
+import { useCourseReader } from '@/hooks/coursify/useCourseReader';
+import { useTableOfContents } from '@/hooks/coursify/useTableOfContents';
+import { useReaderUI } from '@/hooks/coursify/useReaderUI';
+
+import ReaderHeader from '@/components/coursify/reader/ReaderHeader';
+import ReaderSidebar from '@/components/coursify/reader/ReaderSidebar';
+import CourseOverview from '@/components/coursify/reader/CourseOverview';
+import MarkdownRenderer from '@/components/coursify/reader/MarkdownRenderer';
+import TableOfContents from '@/components/coursify/reader/TableOfContents';
+import ReaderNavigation from '@/components/coursify/reader/ReaderNavigation';
 
 const pacifico = Pacifico({
   weight: '400',
@@ -44,12 +37,6 @@ const lora = Lora({
   display: 'swap',
 });
 
-const DIFFICULTY_COLORS = {
-  beginner: 'bg-emerald-100 text-emerald-700',
-  intermediate: 'bg-amber-100 text-amber-700',
-  advanced: 'bg-red-100 text-red-700',
-};
-
 const RESOURCE_ICONS = {
   video: '▶',
   article: '📄',
@@ -59,51 +46,43 @@ const RESOURCE_ICONS = {
 
 export default function PublicCourseReaderPage({ params }) {
   const { slug } = use(params);
-  const id = slug;
   const router = useRouter();
-
-  const [course, setCourse] = useState(null);
-  const [sections, setSections] = useState([]);
-  const [modules, setModules] = useState([]);
-  const [activeSection, setActiveSection] = useState(null);
-  const [showOverview, setShowOverview] = useState(true);
-  const [visited, setVisited] = useState(new Set());
-  const [isLoading, setIsLoading] = useState(true);
-  const [notFound, setNotFound] = useState(false);
-  const [sidebarOpen, setSidebarOpen] = useState(false);
-  const [expandedModules, setExpandedModules] = useState(new Set());
-
-  // TOC state — must be before early returns
-  const [headings, setHeadings] = useState([]);
-  const [activeHeading, setActiveHeading] = useState(null);
-  const [tocOpen, setTocOpen] = useState(true);
   const contentRef = useRef(null);
 
-  const fetchCourse = useCallback(async () => {
-    try {
-      const res = await fetch(`/api/coursify/public/courses/${id}`);
-      const data = await res.json();
-      if (data.success) {
-        setCourse(data.course);
-        setSections(data.sections);
-        setModules(data.modules || []);
-        // Expand all modules by default
-        if (data.modules) {
-          setExpandedModules(new Set(data.modules.map((m) => m._id)));
-        }
-      } else {
-        setNotFound(true);
-      }
-    } catch {
-      setNotFound(true);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [id]);
+  const {
+    course,
+    sections,
+    modules,
+    activeSection,
+    showOverview,
+    visited,
+    isLoading,
+    error,
+    fetchCourse,
+    navigateTo,
+    showOverviewPage,
+  } = useCourseReader(slug, true);
+
+  const {
+    sidebarOpen,
+    tocOpen,
+    expandedModules,
+    setExpandedModules,
+    toggleSidebar,
+    toggleToc,
+    toggleModule,
+  } = useReaderUI(modules);
+
+  const currentSection = sections.find((s) => s._id === activeSection);
+  const { headings, activeHeading } = useTableOfContents(currentSection?.content, contentRef);
 
   useEffect(() => {
-    fetchCourse();
-  }, [fetchCourse]);
+    fetchCourse().then((data) => {
+      if (data?.modules) {
+        setExpandedModules(new Set(data.modules.map((m) => m._id)));
+      }
+    });
+  }, [fetchCourse, setExpandedModules]);
 
   // Expand module when section changes
   useEffect(() => {
@@ -118,7 +97,7 @@ export default function PublicCourseReaderPage({ params }) {
         });
       }
     }
-  }, [activeSection, sections]);
+  }, [activeSection, sections, setExpandedModules]);
 
   // Expose current section context to ChatbotWidget via window global
   useEffect(() => {
@@ -135,53 +114,6 @@ export default function PublicCourseReaderPage({ params }) {
       delete window.__coursifyCtx;
     };
   }, [activeSection, sections]);
-
-  // currentSection computed here so TOC effects can access it unconditionally
-  const currentSection = sections.find((s) => s._id === activeSection);
-  const currentIndex = sections.findIndex((s) => s._id === activeSection);
-
-  // TOC: extract headings from current section content
-  useEffect(() => {
-    if (!currentSection?.content) {
-      setHeadings([]);
-      setActiveHeading(null);
-      return;
-    }
-    const lines = currentSection.content.split('\n');
-    const extracted = [];
-    for (const line of lines) {
-      const h2Match = line.match(/^##\s+(.+)$/);
-      const h3Match = line.match(/^###\s+(.+)$/);
-      if (h2Match) {
-        const text = h2Match[1].trim();
-        extracted.push({ level: 2, text, slug: text.toLowerCase().replace(/[^a-z0-9]+/g, '-') });
-      } else if (h3Match) {
-        const text = h3Match[1].trim();
-        extracted.push({ level: 3, text, slug: text.toLowerCase().replace(/[^a-z0-9]+/g, '-') });
-      }
-    }
-    setHeadings(extracted);
-    setActiveHeading(extracted[0]?.text || null);
-  }, [currentSection?.content]);
-
-  // TOC: scroll-spy
-  useEffect(() => {
-    if (headings.length === 0 || !contentRef.current) return;
-    const observer = new IntersectionObserver(
-      (entries) => {
-        for (const entry of entries) {
-          if (entry.isIntersecting) {
-            const text = entry.target.getAttribute('data-heading');
-            if (text) setActiveHeading(text);
-          }
-        }
-      },
-      { rootMargin: '-80px 0px -60% 0px', threshold: 0 }
-    );
-    const headingEls = contentRef.current.querySelectorAll('[data-heading]');
-    headingEls.forEach((el) => observer.observe(el));
-    return () => observer.disconnect();
-  }, [headings]);
 
   if (isLoading) {
     return (
@@ -202,7 +134,7 @@ export default function PublicCourseReaderPage({ params }) {
     );
   }
 
-  if (notFound || !course) {
+  if (error || !course) {
     return (
       <div
         className={`min-h-screen bg-[#fcfbf5] ${pacifico.variable} ${nunito.variable} font-[family-name:var(--font-sans)] flex items-center justify-center`}
@@ -226,223 +158,45 @@ export default function PublicCourseReaderPage({ params }) {
     );
   }
 
-  const navigateTo = (sectionId) => {
-    if (activeSection) setVisited((v) => new Set(v).add(activeSection));
-    setShowOverview(false);
-    setActiveSection(sectionId);
-    setSidebarOpen(false);
+  const handleNavigate = (id) => {
+    navigateTo(id);
+    toggleSidebar(false);
   };
 
-  const showOverviewPage = () => {
-    setShowOverview(true);
-    setActiveSection(null);
-    setSidebarOpen(false);
+  const handleShowOverview = () => {
+    showOverviewPage();
+    toggleSidebar(false);
   };
-
-  const toggleModule = (moduleId) => {
-    setExpandedModules((prev) => {
-      const next = new Set(prev);
-      if (next.has(moduleId)) next.delete(moduleId);
-      else next.add(moduleId);
-      return next;
-    });
-  };
-
-  const progressPct = sections.length > 0 ? ((currentIndex + 1) / sections.length) * 100 : 0;
 
   return (
     <div
       className={`h-screen bg-[#fcfbf5] font-[family-name:var(--font-sans)] text-[#1e3a34] flex flex-col ${pacifico.variable} ${nunito.variable} ${lora.variable}`}
     >
-      {/* ── Header ── */}
-      <header className="sticky top-0 z-40 bg-white border-b border-[#e5e3d8] px-4 lg:px-6 py-3 flex items-center justify-between gap-3 shrink-0">
-        <div className="flex items-center gap-2 min-w-0">
-          <button
-            onClick={() => router.push('/coursify')}
-            className="p-1.5 hover:bg-[#f0f5f2] rounded-full transition-colors shrink-0"
-          >
-            <ArrowLeft className="w-4 h-4" />
-          </button>
-          <img
-            src="/images/apps/coursify.png"
-            alt="Coursify"
-            className="h-6 w-6 rounded-md object-contain shrink-0 hidden sm:block"
-          />
-          <span className="font-[family-name:var(--font-logo)] text-lg text-[#1f644e] shrink-0 hidden sm:block">
-            Coursify
-          </span>
-          <ChevronRight className="w-3.5 h-3.5 text-[#e5e3d8] shrink-0 hidden sm:block" />
-          <h1 className="font-bold text-[#1e3a34] text-sm lg:text-base truncate">{course.title}</h1>
-        </div>
-
-        <div className="flex items-center gap-2 shrink-0">
-          <div className="hidden sm:flex items-center gap-2">
-            <span
-              className={`px-2 py-0.5 rounded-full text-[10px] font-bold capitalize ${DIFFICULTY_COLORS[course.difficulty] || DIFFICULTY_COLORS.beginner}`}
-            >
-              {course.difficulty}
-            </span>
-            {course.estimatedDuration && (
-              <span className="flex items-center gap-1 text-[10px] text-[#7c8e88] font-bold">
-                <Clock className="w-3 h-3" />
-                {course.estimatedDuration}
-              </span>
-            )}
-            {showOverview ? (
-              <span className="flex items-center gap-1 text-[10px] text-[#1f644e] font-bold">
-                <BookOpen className="w-3 h-3" />
-                Overview
-              </span>
-            ) : null}
-          </div>
-          <button
-            onClick={() => setSidebarOpen((v) => !v)}
-            className="p-1.5 rounded-lg border border-[#e5e3d8] bg-white text-[#7c8e88] lg:hidden"
-          >
-            <Menu className="w-4 h-4" />
-          </button>
-        </div>
-      </header>
+      <ReaderHeader
+        course={course}
+        showOverview={showOverview}
+        onBack={() => router.push('/coursify')}
+        onToggleSidebar={toggleSidebar}
+      />
 
       <div className="flex flex-1 min-h-0">
-        {/* Mobile overlay */}
-        {sidebarOpen && (
-          <div
-            className="fixed inset-0 bg-black/40 z-30 lg:hidden"
-            onClick={() => setSidebarOpen(false)}
-          />
-        )}
+        <ReaderSidebar
+          course={course}
+          sections={sections}
+          modules={modules}
+          activeSection={activeSection}
+          showOverview={showOverview}
+          visited={visited}
+          sidebarOpen={sidebarOpen}
+          expandedModules={expandedModules}
+          onToggleSidebar={toggleSidebar}
+          onNavigate={handleNavigate}
+          onShowOverview={handleShowOverview}
+          onToggleModule={toggleModule}
+        />
 
-        {/* ── Sidebar ── */}
-        <aside
-          className={`fixed lg:static inset-y-0 left-0 z-40 w-72 shrink-0 bg-white border-r border-[#e5e3d8] flex flex-col transition-transform duration-300 lg:translate-x-0 ${
-            sidebarOpen ? 'translate-x-0' : '-translate-x-full'
-          }`}
-        >
-          {/* Mobile drawer header */}
-          <div className="lg:hidden flex items-center justify-between px-4 py-3 border-b border-[#e5e3d8]">
-            <span className="font-bold text-sm text-[#1e3a34]">Sections</span>
-            <button
-              onClick={() => setSidebarOpen(false)}
-              className="p-1.5 hover:bg-[#f0f5f2] rounded-lg transition-colors"
-            >
-              <X className="w-4 h-4 text-[#7c8e88]" />
-            </button>
-          </div>
-
-          {/* Sidebar header: title + progress */}
-          <div className="px-4 py-3 border-b border-[#e5e3d8] space-y-2.5">
-            <p className="font-bold text-xs text-[#1e3a34] line-clamp-2 leading-snug hidden lg:block">
-              {course.title}
-            </p>
-          </div>
-
-          {/* Section list */}
-          <div className="flex-1 overflow-y-auto p-2">
-            {/* Overview button */}
-            <button
-              onClick={showOverviewPage}
-              className={`w-full text-left px-3 py-2.5 rounded-xl mb-1 transition-colors flex items-center gap-2.5 ${
-                showOverview ? 'bg-[#1f644e] text-white' : 'text-[#1e3a34] hover:bg-[#f0f5f2]'
-              }`}
-            >
-              <BookOpen
-                className={`w-4 h-4 shrink-0 ${showOverview ? 'text-white' : 'text-[#7c8e88]'}`}
-              />
-              <span className="text-xs font-bold truncate">Overview</span>
-            </button>
-            {modules.length > 0 ? (
-              <>
-                {modules.map((mod, modIdx) => {
-                  const modSections = sections.filter((s) => s.moduleId === mod._id);
-                  const doneSections = modSections.filter((s) => visited.has(s._id)).length;
-                  const isExpanded = expandedModules.has(mod._id);
-
-                  return (
-                    <div key={mod._id} className="mb-1">
-                      {/* Module header */}
-                      <button
-                        onClick={() => toggleModule(mod._id)}
-                        className="w-full flex items-center gap-2 px-2 py-2 mt-1 hover:bg-[#f0f5f2] rounded-lg transition-colors group"
-                      >
-                        <span className="w-5 h-5 rounded-md bg-[#1f644e]/10 text-[#1f644e] text-[10px] font-bold flex items-center justify-center shrink-0">
-                          {modIdx + 1}
-                        </span>
-                        <span className="text-[11px] font-bold text-[#1e3a34] truncate flex-1 text-left">
-                          {mod.title}
-                        </span>
-                        <div className="flex items-center gap-1.5 shrink-0">
-                          <span className="text-[10px] font-bold text-[#b0bfbb]">
-                            {doneSections}/{modSections.length}
-                          </span>
-                          <ChevronRight
-                            className={`w-3.5 h-3.5 text-[#b0bfbb] transition-transform duration-200 ${isExpanded ? 'rotate-90' : ''}`}
-                          />
-                        </div>
-                      </button>
-
-                      {isExpanded && (
-                        <div className="ml-2 pl-3 border-l-2 border-[#f0f5f2] space-y-0.5 mt-0.5">
-                          {modSections.map((section) => (
-                            <SidebarSectionBtn
-                              key={section._id}
-                              section={section}
-                              active={activeSection === section._id}
-                              done={visited.has(section._id)}
-                              onClick={() => navigateTo(section._id)}
-                            />
-                          ))}
-                          {modSections.length === 0 && (
-                            <p className="text-[10px] text-[#b0bfbb] px-3 py-1 italic">
-                              No sections yet
-                            </p>
-                          )}
-                        </div>
-                      )}
-                    </div>
-                  );
-                })}
-                {sections.filter((s) => !s.moduleId).length > 0 && (
-                  <div className="mb-1 mt-2">
-                    <div className="flex items-center gap-2 px-2 py-1.5">
-                      <span className="text-[10px] font-bold uppercase tracking-wider text-[#b0bfbb]">
-                        More
-                      </span>
-                    </div>
-                    <div className="space-y-0.5">
-                      {sections
-                        .filter((s) => !s.moduleId)
-                        .map((section) => (
-                          <SidebarSectionBtn
-                            key={section._id}
-                            section={section}
-                            active={activeSection === section._id}
-                            done={visited.has(section._id)}
-                            onClick={() => navigateTo(section._id)}
-                          />
-                        ))}
-                    </div>
-                  </div>
-                )}
-              </>
-            ) : (
-              sections.map((section) => (
-                <SidebarSectionBtn
-                  key={section._id}
-                  section={section}
-                  active={activeSection === section._id}
-                  done={visited.has(section._id)}
-                  onClick={() => navigateTo(section._id)}
-                />
-              ))
-            )}
-          </div>
-        </aside>
-
-        {/* ── Main Content ─ */}
         <main className="flex-1 overflow-y-auto min-w-0">
           <div className="flex min-h-0">
-            {/* Article */}
             <article className="flex-1 max-w-3xl mx-auto px-4 lg:px-10 py-8" ref={contentRef}>
               {!currentSection && !showOverview ? (
                 <div className="flex flex-col items-center justify-center py-24 text-center">
@@ -453,250 +207,15 @@ export default function PublicCourseReaderPage({ params }) {
                   <p className="text-sm text-[#7c8e88]">Check back soon — content is on the way.</p>
                 </div>
               ) : showOverview ? (
-                <>
-                  {/* Course Overview */}
-                  <div>
-                    {/* Thumbnail */}
-                    {course.thumbnail && (
-                      <div className="w-full aspect-video rounded-2xl overflow-hidden mb-8 border border-[#e5e3d8] shadow-sm">
-                        <img
-                          src={course.thumbnail}
-                          alt={course.title}
-                          className="w-full h-full object-cover"
-                        />
-                      </div>
-                    )}
-
-                    {/* Title + meta */}
-                    <div className="flex flex-wrap items-center gap-2 mb-4">
-                      <span
-                        className={`px-2 py-0.5 rounded-full text-[10px] font-bold capitalize ${DIFFICULTY_COLORS[course.difficulty] || DIFFICULTY_COLORS.beginner}`}
-                      >
-                        {course.difficulty}
-                      </span>
-                      {course.estimatedDuration && (
-                        <span className="flex items-center gap-1 text-[10px] font-bold text-[#7c8e88]">
-                          <Clock className="w-3 h-3" />
-                          {course.estimatedDuration}
-                        </span>
-                      )}
-                      {sections.length > 0 && (
-                        <span className="flex items-center gap-1 text-[10px] font-bold text-[#7c8e88]">
-                          <Layers className="w-3 h-3" />
-                          {sections.length} section{sections.length !== 1 ? 's' : ''}
-                        </span>
-                      )}
-                    </div>
-
-                    {/* Description */}
-                    {course.description && (
-                      <p className="text-base text-[#7c8e88] leading-relaxed mb-8">
-                        {course.description}
-                      </p>
-                    )}
-
-                    {/* What you'll learn */}
-                    {course.learningObjectives?.length > 0 && (
-                      <div className="mb-8">
-                        <h3 className="text-sm font-bold text-[#1e3a34] mb-3">
-                          What you&apos;ll learn
-                        </h3>
-                        <ul className="space-y-2">
-                          {course.learningObjectives.map((obj, i) => (
-                            <li key={i} className="flex items-start gap-2.5 text-sm text-[#1e3a34]">
-                              <Check className="w-4 h-4 text-[#1f644e] shrink-0 mt-0.5" />
-                              <span>{obj}</span>
-                            </li>
-                          ))}
-                        </ul>
-                      </div>
-                    )}
-
-                    {/* Prerequisites */}
-                    {course.prerequisites?.length > 0 && (
-                      <div className="mb-8">
-                        <h3 className="text-sm font-bold text-[#1e3a34] mb-3">Prerequisites</h3>
-                        <ul className="space-y-2">
-                          {course.prerequisites.map((pre, i) => (
-                            <li key={i} className="flex items-start gap-2.5 text-sm text-[#7c8e88]">
-                              <span className="w-1.5 h-1.5 rounded-full bg-[#e5e3d8] shrink-0 mt-1.5" />
-                              <span>{pre}</span>
-                            </li>
-                          ))}
-                        </ul>
-                      </div>
-                    )}
-
-                    {/* Section outline */}
-                    {sections.length > 0 && (
-                      <div className="mb-8">
-                        <h3 className="text-sm font-bold text-[#1e3a34] mb-3">
-                          {modules.length > 0 ? 'Course outline' : 'Sections'}
-                        </h3>
-                        <div className="space-y-1">
-                          {modules.length > 0 ? (
-                            <>
-                              {modules.map((mod, modIdx) => {
-                                const modSections = sections.filter((s) => s.moduleId === mod._id);
-                                return (
-                                  <div key={mod._id} className="mb-3">
-                                    <p className="text-[10px] font-bold uppercase tracking-wider text-[#7c8e88] mb-1 px-2">
-                                      {modIdx + 1}. {mod.title}
-                                    </p>
-                                    <div className="ml-4 space-y-0.5">
-                                      {modSections.map((section) => (
-                                        <button
-                                          key={section._id}
-                                          onClick={() => navigateTo(section._id)}
-                                          className="w-full text-left px-3 py-2 rounded-lg text-xs font-bold text-[#1e3a34] hover:bg-[#f0f5f2] transition-colors flex items-center gap-2"
-                                        >
-                                          {section.title}
-                                        </button>
-                                      ))}
-                                    </div>
-                                  </div>
-                                );
-                              })}
-                              {sections.filter((s) => !s.moduleId).length > 0 && (
-                                <div className="mb-3">
-                                  <p className="text-[10px] font-bold uppercase tracking-wider text-[#7c8e88] mb-1 px-2">
-                                    More
-                                  </p>
-                                  <div className="ml-4 space-y-0.5">
-                                    {sections
-                                      .filter((s) => !s.moduleId)
-                                      .map((section) => (
-                                        <button
-                                          key={section._id}
-                                          onClick={() => navigateTo(section._id)}
-                                          className="w-full text-left px-3 py-2 rounded-lg text-xs font-bold text-[#1e3a34] hover:bg-[#f0f5f2] transition-colors flex items-center gap-2"
-                                        >
-                                          {section.title}
-                                        </button>
-                                      ))}
-                                  </div>
-                                </div>
-                              )}
-                            </>
-                          ) : (
-                            sections.map((section) => (
-                              <button
-                                key={section._id}
-                                onClick={() => navigateTo(section._id)}
-                                className="w-full text-left px-3 py-2 rounded-lg text-xs font-bold text-[#1e3a34] hover:bg-[#f0f5f2] transition-colors flex items-center gap-2"
-                              >
-                                {section.title}
-                              </button>
-                            ))
-                          )}
-                        </div>
-                      </div>
-                    )}
-
-                    {/* Start Course button */}
-                    {sections.length > 0 && (
-                      <button
-                        onClick={() => navigateTo(sections[0]._id)}
-                        className="w-full flex items-center justify-center gap-2 px-6 py-3 bg-[#1f644e] text-white rounded-xl text-sm font-bold hover:bg-[#17503e] transition-colors"
-                      >
-                        Start learning <ChevronRight className="w-4 h-4" />
-                      </button>
-                    )}
-                  </div>
-                </>
+                <CourseOverview
+                  course={course}
+                  sections={sections}
+                  modules={modules}
+                  onNavigate={handleNavigate}
+                />
               ) : (
                 <>
-                  {/* Markdown content */}
-                  {currentSection.content ? (
-                    <div className="coursify-md prose prose-sm max-w-none font-[family-name:var(--font-lora)] prose-headings:font-bold prose-headings:text-[#1e3a34] prose-p:text-[#1e3a34] prose-p:leading-relaxed prose-code:bg-[#f0f5f2] prose-code:rounded prose-code:px-1 prose-code:text-[#1f644e] prose-pre:bg-[#1e3a34] prose-pre:rounded-xl prose-blockquote:border-[#1f644e] prose-a:text-[#1f644e] prose-li:text-[#1e3a34] prose-strong:text-[#1e3a34] prose-table:text-sm">
-                      <ReactMarkdown
-                        remarkPlugins={[remarkGfm]}
-                        rehypePlugins={[rehypeRaw]}
-                        components={{
-                          h2({ children, ...props }) {
-                            const text = typeof children === 'string' ? children : '';
-                            const slug = text.toLowerCase().replace(/[^a-z0-9]+/g, '-');
-                            return (
-                              <h2 id={slug} data-heading={text} className="scroll-mt-20" {...props}>
-                                {children}
-                              </h2>
-                            );
-                          },
-                          h3({ children, ...props }) {
-                            const text = typeof children === 'string' ? children : '';
-                            const slug = text.toLowerCase().replace(/[^a-z0-9]+/g, '-');
-                            return (
-                              <h3 id={slug} data-heading={text} className="scroll-mt-20" {...props}>
-                                {children}
-                              </h3>
-                            );
-                          },
-                          table({ children }) {
-                            return (
-                              <div className="overflow-x-auto my-7 rounded-xl border border-[#e5e3d8]">
-                                <table className="w-full border-collapse text-sm">{children}</table>
-                              </div>
-                            );
-                          },
-                          code({ node, className, children, ...props }) {
-                            const match = /language-(\w+)/.exec(className || '');
-                            const isBlock = String(children).includes('\n');
-
-                            if (isBlock && match) {
-                              return (
-                                <SyntaxHighlighter
-                                  style={oneDark}
-                                  language={match[1]}
-                                  PreTag="div"
-                                  customStyle={{
-                                    borderRadius: '0.75rem',
-                                    fontSize: '0.82rem',
-                                    margin: '0.75em 0',
-                                    padding: '0.6em 0.9em',
-                                  }}
-                                  showLineNumbers
-                                  {...props}
-                                >
-                                  {String(children).replace(/\n$/, '')}
-                                </SyntaxHighlighter>
-                              );
-                            }
-                            if (isBlock) {
-                              return (
-                                <div
-                                  className="rounded-xl overflow-hidden my-3 w-full flex justify-center"
-                                  style={{ background: '#18181b' }}
-                                >
-                                  <pre
-                                    className="overflow-x-auto p-4 text-[0.82rem] leading-relaxed font-mono whitespace-pre"
-                                    style={{ background: 'transparent' }}
-                                  >
-                                    <code className="font-mono" style={{ color: '#e4e4e7' }}>
-                                      {children}
-                                    </code>
-                                  </pre>
-                                </div>
-                              );
-                            }
-                            return (
-                              <code
-                                className="bg-[#f0f5f2] text-[#1f644e] rounded px-1.5 py-0.5 text-[0.82em] font-mono font-semibold"
-                                {...props}
-                              >
-                                {children}
-                              </code>
-                            );
-                          },
-                        }}
-                      >
-                        {currentSection.content}
-                      </ReactMarkdown>
-                    </div>
-                  ) : (
-                    <div className="py-16 text-center">
-                      <p className="text-sm text-[#7c8e88]">No content yet for this section.</p>
-                    </div>
-                  )}
+                  <MarkdownRenderer content={currentSection.content} />
 
                   {/* Resources */}
                   {currentSection.resources?.length > 0 && (
@@ -727,136 +246,24 @@ export default function PublicCourseReaderPage({ params }) {
                     </div>
                   )}
 
-                  {/* Prev / Next navigation */}
-                  <SectionNav
+                  <ReaderNavigation
                     sections={sections}
                     activeSection={activeSection}
-                    onNavigate={navigateTo}
+                    onNavigate={handleNavigate}
                   />
                 </>
               )}
             </article>
 
-            {/* ── TOC Sidebar (desktop only) ── */}
-            {headings.length > 0 && (
-              <aside className="hidden lg:block shrink-0 sticky top-0 self-start max-h-screen overflow-y-auto py-8 pr-2 pl-2">
-                <div className="flex items-center justify-between gap-2 mb-2">
-                  <div className="flex items-center gap-1.5 text-[#7c8e88]">
-                    <List className="w-3.5 h-3.5" />
-                    <span className="text-[10px] font-bold uppercase tracking-wider">
-                      On this page
-                    </span>
-                  </div>
-                  <button
-                    onClick={() => setTocOpen((v) => !v)}
-                    className="p-0.5 rounded text-[#7c8e88] hover:text-[#1e3a34] hover:bg-[#f0f5f2] transition-colors"
-                    title={tocOpen ? 'Collapse' : 'Expand'}
-                  >
-                    <ChevronRight
-                      className={`w-3.5 h-3.5 transition-transform duration-200 ${tocOpen ? 'rotate-90' : ''}`}
-                    />
-                  </button>
-                </div>
-                {tocOpen && (
-                  <nav className="space-y-0.5">
-                    {headings.map((h) => (
-                      <a
-                        key={h.slug}
-                        href={`#${h.slug}`}
-                        onClick={(e) => {
-                          e.preventDefault();
-                          const el = document.getElementById(h.slug);
-                          el?.scrollIntoView({ behavior: 'smooth', block: 'start' });
-                        }}
-                        className={`block text-xs leading-snug transition-colors py-0.5 ${
-                          h.level === 3 ? 'pl-4' : ''
-                        } ${
-                          activeHeading === h.text
-                            ? 'text-[#1f644e] font-bold'
-                            : 'text-[#7c8e88] hover:text-[#1e3a34]'
-                        }`}
-                      >
-                        {h.text}
-                      </a>
-                    ))}
-                  </nav>
-                )}
-              </aside>
-            )}
+            <TableOfContents
+              headings={headings}
+              activeHeading={activeHeading}
+              tocOpen={tocOpen}
+              onToggleToc={toggleToc}
+            />
           </div>
         </main>
       </div>
-    </div>
-  );
-}
-
-function SidebarSectionBtn({ section, active, done, onClick }) {
-  return (
-    <button
-      onClick={onClick}
-      className={`w-full text-left px-3 py-2 rounded-lg transition-colors flex items-center gap-2.5 ${
-        active ? 'bg-[#1f644e] text-white shadow-sm' : 'text-[#1e3a34] hover:bg-[#f0f5f2]'
-      }`}
-    >
-      <span
-        className={`text-xs font-semibold truncate flex-1 leading-snug ${
-          done && !active ? 'text-[#7c8e88]' : ''
-        }`}
-      >
-        {section.title}
-      </span>
-    </button>
-  );
-}
-
-function SectionNav({ sections, activeSection, onNavigate }) {
-  const idx = sections.findIndex((s) => s._id === activeSection);
-  const prev = idx > 0 ? sections[idx - 1] : null;
-  const next = idx < sections.length - 1 ? sections[idx + 1] : null;
-
-  if (!prev && !next) return null;
-
-  return (
-    <div className="mt-10 pt-6 border-t border-[#e5e3d8] flex items-center justify-between gap-4">
-      {prev ? (
-        <button
-          onClick={() => onNavigate(prev._id)}
-          className="flex items-center gap-2 text-left group"
-        >
-          <div className="p-1.5 rounded-lg border border-[#e5e3d8] text-[#7c8e88] group-hover:border-[#1f644e] group-hover:text-[#1f644e] transition-colors">
-            <ArrowLeft className="w-3.5 h-3.5" />
-          </div>
-          <div className="min-w-0">
-            <p className="text-[10px] text-[#7c8e88] font-bold uppercase tracking-wider">
-              Previous
-            </p>
-            <p className="text-xs font-bold text-[#1e3a34] truncate max-w-[160px] group-hover:text-[#1f644e] transition-colors">
-              {prev.title}
-            </p>
-          </div>
-        </button>
-      ) : (
-        <div />
-      )}
-
-      {next && (
-        <button
-          onClick={() => onNavigate(next._id)}
-          className="flex items-center gap-2 text-right group ml-auto"
-        >
-          <div className="min-w-0">
-            <p className="text-[10px] text-[#7c8e88] font-bold uppercase tracking-wider text-right">
-              Next
-            </p>
-            <p className="text-xs font-bold text-[#1e3a34] truncate max-w-[160px] group-hover:text-[#1f644e] transition-colors">
-              {next.title}
-            </p>
-          </div>
-          <div className="p-1.5 rounded-lg border border-[#e5e3d8] text-[#7c8e88] group-hover:border-[#1f644e] group-hover:text-[#1f644e] transition-colors">
-            <ChevronRight className="w-3.5 h-3.5" />
-          </div>
-        </button>
-      )}
     </div>
   );
 }
